% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{alignment_heterozygosity}
\alias{alignment_heterozygosity}
\title{calculated positional heterozygosity along an alignment}
\usage{
alignment_heterozygosity(alignment)
}
\arguments{
\item{alignment}{fasta file location}
}
\value{
dataframe of postiion and heterozygosity
}
\description{
calculated positional heterozygosity along an alignment
}
