% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions_coarse_parallell.R
\name{get_regions_coarse_parallel}
\alias{get_regions_coarse_parallel}
\title{For a given sequence alignment, fasta format, file name. Return the start and stop positions relative to a reference sequence. Quick and Dirty method only.}
\usage{
get_regions_coarse_parallel(
  alignment = "all_raw_best_msa_man3b.fasta",
  ref.file = "ref/NC_006273.2.fasta",
  ref.pattern = "Merlin",
  run_width = 200,
  run_java = "java",
  run_model = "AIC"
)
}
\arguments{
\item{alignment}{Multiple Sequence Alignment fasta file location}

\item{ref.file}{Fasta file containing Reference genome}

\item{ref.pattern}{}

\item{run_width}{how many base pairs should each genome scan be}

\item{run_java}{command to run java executable, equired openjdk 13 or equivilent}

\item{run_model}{model selection either 'AIC' or 'BIC. Defaut is AIC}
}
\value{
vector representing start and stop positions relative to reference
}
\description{
For a given sequence alignment, fasta format, file name. Return the start and stop positions relative to a reference sequence. Quick and Dirty method only.
}
