% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions_parallell.R
\name{get_regions_parallel}
\alias{get_regions_parallel}
\title{For a given sequence alignment, fasta format, file name. Return the start and stop positions relative to a reference sequence}
\usage{
get_regions_parallel(
  alignment = system.file("extdata/cmv_msa.fasta", package = "hmmcluster"),
  ref.file = system.file("ref/NC_006273.2.fasta", package = "hmmcluster"),
  ref.pattern = "Merlin",
  run_width = 200,
  run_java = "java",
  run_model = "AIC",
  run_cores = 1
)
}
\arguments{
\item{alignment}{Multiple Sequence Alignment fasta file location}

\item{ref.file}{Fasta file containing Reference genome}

\item{ref.pattern}{Unique string to identift reference sequence in alignment labels}

\item{run_width}{how many base pairs should each genome scan be}

\item{run_java}{command to run java executable, equired openjdk 13 or equivilent}

\item{run_model}{model selection either 'AIC' or 'BIC. Defaut is AIC}

\item{run_cores}{how many cores for parallell processes to use}
}
\value{
vector representing start and stop positions relative to reference
}
\description{
For a given sequence alignment, fasta format, file name. Return the start and stop positions relative to a reference sequence
}
