% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relative_ref_pos.R
\name{get_relative_ref_pos}
\alias{get_relative_ref_pos}
\title{Return the start and stop positions relative to my reference sequence}
\usage{
get_relative_ref_pos(
  alignment,
  ref.file = system.file("ref/NC_006273.2.fasta", package = "hmmcluster"),
  ref.pattern = "Merlin"
)
}
\arguments{
\item{alignment}{Multiple Sequence Alignment as dnabin}

\item{ref.file}{Fasta file containing Reference genome}

\item{ref.pattern}{Unique string to identift reference sequence in alignment labels}
}
\value{
vector representing start and stop positions relative to reference
}
\description{
Return the start and stop positions relative to my reference sequence
}
