% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relative_ref_pos.R
\name{get_relative_ref_pos_safe}
\alias{get_relative_ref_pos_safe}
\title{Safely, escape if reference sequence has been corrupted. Return the start and stop positions relative to my reference sequence}
\usage{
get_relative_ref_pos_safe(
  alignment,
  ref.file = system.file("ref/NC_006273.2.fasta", package = "hmmcluster"),
  ref.pattern = "Merlin"
)
}
\arguments{
\item{alignment}{Multiple Sequence Alignment fasta file location}

\item{ref.file}{Fasta file containing Reference genome}

\item{ref.pattern}{}
}
\value{
vector representing start and stop positions relative to reference
}
\description{
Safely, escape if reference sequence has been corrupted. Return the start and stop positions relative to my reference sequence
}
