% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{hzyg}
\alias{hzyg}
\title{calculates heterozygosity for a position, in an alignment (ape package matrix)}
\usage{
hzyg(i, seq)
}
\arguments{
\item{i}{position}

\item{seq}{alignement as character matrix}
}
\value{
vector os position and heterozygocuty
}
\description{
calculates heterozygosity for a position, in an alignment (ape package matrix)
}
